// frontend/src/shared/api/generated/models/announcementListItemResponse.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * CodigoActivo.API
 * OpenAPI spec version: 1.0
 */

export interface AnnouncementListItemResponse {
  id?: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  subtitle?: string | null;
  createdAt?: string;
  /** @nullable */
  updatedAt?: string | null;
  createdBy?: string;
  /** @nullable */
  updatedBy?: string | null;
  thumbnailId?: string;
  featured?: boolean;
}